+++ pkg/serve/settings.go
@@ -1,0 +1,15 @@
+package serve
+
+type Setting struct {
+	Type  SettingType
+	Value interface{}
+}
+
+type SettingType uint
+
+const (
+	Directory SettingType = iota
+	Verbose
+	IPAddress
+	Port
+)